MXS-4973 Don't create test-admin user on local servers
Would require unix socket access. Assume that local servers
always have this user.

+++ system-test/maxtest/src/mariadb_nodes.cc
@@ -349,18 +349,37 @@ bool MariaDBCluster::create_base_users(int node)
 {
     using mxt::MariaDBServer;
 
-    // Create the basic test admin user with ssh as the backend may not accept external connections.
-    // The sql-command given to ssh must escape double quotes.
+    bool test_admin_user_ok = false;
     auto vm = this->node(node);
-    string drop_query = mxb::string_printf(R"(drop user \"%s\";)", admin_user.c_str());
-    vm->run_sql_query(drop_query);
-    string create_query = mxb::string_printf(
-        R"(create user \"%s\" identified by \"%s\"; grant all on *.* to \"%s\" with grant option;)",
-        admin_user.c_str(), admin_pw.c_str(), admin_user.c_str());
-    auto res = vm->run_sql_query(create_query);
+
+    if (vm->is_remote())
+    {
+        // Create the basic test admin user with ssh as the backend may not accept external connections.
+        // The sql-command given to ssh must escape double quotes.
+        string drop_query = mxb::string_printf(R"(drop user \"%s\";)", admin_user.c_str());
+        vm->run_sql_query(drop_query);
+        string create_query = mxb::string_printf(
+            R"(create user \"%s\" identified by \"%s\"; grant all on *.* to \"%s\" with grant option;)",
+            admin_user.c_str(), admin_pw.c_str(), admin_user.c_str());
+        auto res = vm->run_sql_query(create_query);
+        if (res.rc == 0)
+        {
+            test_admin_user_ok = true;
+        }
+        else
+        {
+            logger().log_msgf("Command '%s' failed on cluster '%s' node %i. Return value: %i, %s.",
+                              create_query.c_str(), name().c_str(), node, res.rc, res.output.c_str());
+        }
+    }
+    else
+    {
+        // If server is running locally, assume the test admin user is always there.
+        test_admin_user_ok = true;
+    }
 
     bool rval = false;
-    if (res.rc == 0)
+    if (test_admin_user_ok)
     {
         auto be = backend(node);
         be->update_status();
@@ -398,11 +417,7 @@ bool MariaDBCluster::create_base_users(int node)
                               name().c_str(), node);
         }
     }
-    else
-    {
-        logger().log_msgf("Command '%s' failed on cluster '%s' node %i. Return value: %i, %s.",
-                          create_query.c_str(), name().c_str(), node, res.rc, res.output.c_str());
-    }
+
     return rval;
 }
